Add new home page, router, and private route component
